+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 ## 1.47.1 (Unreleased)
+
+ENHANCEMENTS:
+
+* Resource: `tencentcloud_clb_listener` support configure HTTP health check for TCP listener([#539](https://github.com/tencentcloudstack/terraform-provider-tencentcloud/issues/539)).
+* Data Source: `tencentcloud_clb_listeners` support getting HTTP health check config for TCP listener.
+
 ## 1.47.0 (November 13, 2020)
 
 ENHANCEMENTS: